Change 'solenoid to set' offset for K270
Optimal sheduling wrt cam position asks to set solenoid
9 and not 6 needles away from the carriage center position

This fix also corrects sensor positions for KH270 (-3 +3)

+++ lib/components/hallsensor/hallsensor.cpp
@@ -29,7 +29,7 @@ bool HallSensor::isActive() {
   return (_minimum.value != NONE) || (_maximum.value != NONE);
 }
 
-uint8_t HallSensor::getSensorPosition() { return _config->position; }
+int16_t HallSensor::getSensorPosition() { return _config->position; }
 
 int16_t HallSensor::getDetectedPosition() { return _detectedPosition; }
